Integration of SQL checks database

+++ src/main/java/com/sparta/coldplay/dto/EmployeeDTO.java
@@ -65,7 +65,7 @@ public class EmployeeDTO {
         return dateOfJoining;
     }
 
-    public Float getSalary() {
+    public int getSalary() {
         return salary;
     }
